[frontend] fix: modify NavigationItem, NavigationMenu
