+++ frontend/stt_diary/src/App.js
@@ -1,25 +1,17 @@
 import './App.css';
-import { useState, useEffect } from 'react';
-import { Routes, Route, useLocation, useNavigate } from 'react-router-dom';
+import { useState } from 'react';
+import { Routes, Route} from 'react-router-dom';
 import DiaryList from './routes/DiaryList';
 import NavigationMenu from './components/NavigationMenu';
 
 
 function App() {
-  const [activePage, setActivePage] = useState('');
-  const location = useLocation();
-  const navigate = useNavigate();
   const [isDropdownOpen, setIsDropdownOpen] = useState(false);
 
   const toggleDropdown = () => {
     setIsDropdownOpen(!isDropdownOpen);
   };
   
-  useEffect(() => {
-    const path = location.pathname;
-    const page = path.split("/").pop();
-    setActivePage(page);
-  }, [location]);
 
   return (
     <div className="App">
@@ -30,11 +22,11 @@ function App() {
             <img src={require('./assets/barIcon.png')} alt=""></img>
           </button>
           <div className={`dropdown-content ${isDropdownOpen ? 'show' : ''}`}>
-            <NavigationMenu activePage={activePage} navigate={navigate} toggleDropdown={toggleDropdown} />
+            <NavigationMenu toggleDropdown={toggleDropdown} />
           </div>
         </div>
         <ul className="navbar">
-          <NavigationMenu activePage={activePage} navigate={navigate} toggleDropdown={toggleDropdown} />
+          <NavigationMenu toggleDropdown={toggleDropdown} />
         </ul>
       </nav>
       <Routes>
