+++ frontend/stt_diary/src/components/NavigationMenu.js
@@ -1,14 +1,18 @@
 
+import React from 'react';
+import { useNavigate } from 'react-router-dom';
 import NavigationItem from "./NavigationItem";
 
-function NavigationMenu({ activePage, navigate, toggleDropdown }) {
-    return (
-      <>
-        <NavigationItem path="" activePage={activePage} onClick={() => {navigate('/'); toggleDropdown();}} title="나의 일기장" />
-        <NavigationItem path="analysis" activePage={activePage} onClick={() => {navigate('/analysis'); toggleDropdown();}} title="나의 감정분석" />
-        <NavigationItem path="login" activePage={activePage} onClick={() => {navigate('/login'); toggleDropdown();}} title="로그아웃" />
-      </>
-    );
+function NavigationMenu({toggleDropdown}) {
+  const navigate = useNavigate();
+  
+  return (
+    <>
+      <NavigationItem path="" title="나의 일기장" navigate={navigate} toggleDropdown={toggleDropdown}/>
+      <NavigationItem path="analysis" title="나의 감정분석" navigate={navigate} toggleDropdown={toggleDropdown} />
+      <NavigationItem path="login" title="로그아웃" navigate={navigate} toggleDropdown={toggleDropdown} />
+    </>
+  );
 }
 
 export default NavigationMenu;
